// src/app/clearing/models/itemcess.ts
import { GlobalVariables } from '../../core/models/globalvariables';
export class ItemCess {
  cess_pkid: string;
  cess_job_id: string;
  cess_itm_id: string;
  cess_srno: number;
  cess_code: string;
  cess_qty: number;
  cess_unit_id: string;
  cess_unit_code: string;
  cess_unit_name: string;
  cess_rate: number;
  cess_ctr: number;
  cess_type:string;
  
  rec_mode: string;
  rec_category: string;
  _globalvariables: GlobalVariables;
}